+++ examples/main/shared/editor/EditorPlugins.jsx
@@ -234,7 +234,7 @@ const videoHandlers = {
   },
 };
 
-export const config = {
+const config = {
   [LINK_PREVIEW_TYPE]: {
     disableEmbed: false,
     fetchMetadata: linkPreviewUtil(authorization),
@@ -256,8 +256,6 @@ export const config = {
     // },
   },
   [GALLERY_TYPE]: {
-    scrollingElement: () =>
-      typeof window !== 'undefined' && document.getElementsByClassName('editor-example')[0],
     // toolbar: {
     //   icons: {
     //     InsertPluginButtonIcon: MyCustomIcon,
@@ -675,4 +673,13 @@ export const config = {
     //   })
     // }
   ],
+};
+
+export const getConfig = (additionalConfig = {}) => {
+  let _config = { ...config };
+  Object.keys(additionalConfig).forEach(key => {
+    _config[key] = { ...(_config[key] || {}), ...(additionalConfig[key] || {}) };
+  });
+
+  return _config;
 };