+++ examples/storybook/stories/Components/ThemesExample.js
@@ -4,7 +4,7 @@ import Palette from './Palette';
 import { RichContentViewerBox } from './StoryParts';
 import { Themes } from '../../src/RceTheme';
 import { wixPalettes } from '../palettesExample';
-import ThemeWrapper from '../../src/ThemeWrapper';
+import { RichContentWrapper } from 'wix-rich-content-wrapper';
 
 export default ({ children }) => {
   return Object.keys(wixPalettes).map((paletteKey, i) => {
@@ -13,11 +13,10 @@ export default ({ children }) => {
       <div key={`palette${i}`}>
         <h3>Palette {i + 1}</h3>
         <Palette palette={palette} />
-
         <RichContentViewerBox preset="blog-preset">
-          <ThemeWrapper theme={Themes.PALETTE} palette={palette}>
+          <RichContentWrapper theme={Themes.PALETTE} palette={palette}>
             {Children.only(children)}
-          </ThemeWrapper>
+          </RichContentWrapper>
         </RichContentViewerBox>
       </div>
     );