+++ packages/plugin-emoji/web/src/toolbar/insert-buttons.jsx
@@ -2,7 +2,7 @@ import { DesktopFlyOutModalStyles } from '../constants';
 import {
   TOOLBARS,
   decorateComponentWithProps,
-  getModalStyles,
+  getBottomToolbarModalStyles,
   DECORATION_MODE,
 } from 'wix-rich-content-editor-common';
 import EmojiPreviewModal from './emojiPreviewModal';
@@ -11,6 +11,7 @@ import EmojiPluginIcon from '../icons/EmojiPluginIcon.svg';
 
 export default ({ helpers, t, settings, getEditorState, setEditorState }) => {
   const icon = settings?.toolbar?.icons?.InsertPluginButtonIcon || EmojiPluginIcon;
+
   return [
     {
       type: 'modal',
@@ -25,24 +26,9 @@ export default ({ helpers, t, settings, getEditorState, setEditorState }) => {
         ...settings,
       }),
       modalStylesFn: ({ buttonRef }) => {
-        const modalStyles = getModalStyles({
+        return getBottomToolbarModalStyles(buttonRef, {
           customStyles: DesktopFlyOutModalStyles,
-          fullScreen: true,
         });
-        const { top, left } = buttonRef.getBoundingClientRect();
-        const modalLeft = left - 288;
-        const isAboveButton = top - 293 > 0;
-        const modalTop = isAboveButton ? top - 293 : top + 30;
-        return {
-          ...modalStyles,
-          content: {
-            ...modalStyles.content,
-            top: modalTop,
-            left: modalLeft,
-            margin: 0,
-            position: 'absolute',
-          },
-        };
       },
       modalDecorations: [
         {