+++ packages/plugin-gallery/web/src/editor.js
@@ -1,0 +1,13 @@
+import { createGalleryPlugin } from './createGalleryPlugin';
+import { GALLERY_TYPE } from './types';
+import { ModalsMap } from './modals';
+import { DEFAULTS } from './constants';
+
+export const pluginGallery = (config = {}) => {
+  return {
+    config: { ...DEFAULTS.config, ...config },
+    type: GALLERY_TYPE,
+    createPlugin: createGalleryPlugin,
+    ModalsMap,
+  };
+};