+++ packages/plugin-gallery/web/src/index.js
@@ -1,3 +1,4 @@
 export { createGalleryPlugin } from './createGalleryPlugin';
 export { GALLERY_TYPE } from './types';
 export { Modals, ModalsMap } from './modals';
+export { pluginGallery } from './editor';