+++ packages/plugin-hashtag/web/src/defaults.js
@@ -1,0 +1,8 @@
+export const DEFAULTS = {
+  config: {
+    createHref: decoratedText => `/search/posts?query=${encodeURIComponent('#')}${decoratedText}`,
+    onClick: event => {
+      event.preventDefault();
+    },
+  },
+};