+++ packages/plugin-hashtag/web/src/editor.js
@@ -1,0 +1,17 @@
+import { createHashtagPlugin } from './createHashtagPlugin';
+import { HASHTAG_TYPE } from './types';
+import { DEFAULTS } from './defaults';
+
+export const pluginHashtag = (config = {}) => {
+  return {
+    config: { ...DEFAULTS.config, ...config },
+    type: HASHTAG_TYPE,
+    createPlugin: createHashtagPlugin,
+    ModalsMap: {},
+    theme: colors => ({
+      hashtag: {
+        color: colors.actionColor,
+      },
+    }),
+  };
+};