+++ packages/plugin-html/web/src/viewer.js
@@ -1,2 +1,13 @@
-export { typeMapper as htmlTypeMapper } from './typeMapper';
-export { HTML_TYPE } from './types';
+import { typeMapper } from './typeMapper';
+import { HTML_TYPE } from './types';
+import { DEFAULTS } from './constants';
+export { typeMapper as htmlTypeMapper, HTML_TYPE };
+
+export const pluginHtml = (config = {}) => {
+  return {
+    config: { ...DEFAULTS.config, ...config },
+    type: HTML_TYPE,
+    typeMapper,
+    decorator: {},
+  };
+};