+++ packages/plugin-video/web/src/video-viewer.jsx
@@ -62,6 +62,7 @@ class VideoViewer extends Component {
   render() {
     this.styles = this.styles || mergeStyles({ styles, theme: this.props.theme });
     const { url, isLoaded, key } = this.state;
+    this.props.setComponentUrl?.(url);
     const props = {
       url,
       onReady: this.onReactPlayerReady,
@@ -92,6 +93,7 @@ VideoViewer.propTypes = {
   theme: PropTypes.object.isRequired,
   disabled: PropTypes.bool,
   disableRightClick: PropTypes.bool,
+  setComponentUrl: PropTypes.func,
 };
 
 VideoViewer.defaultProps = {