+++ packages/wrapper/web/src/index.js
@@ -1,0 +1,3 @@
+import RichContentWrapper from './RichContentWrapper';
+import { wixPalettes } from './palettesExample';
+export { RichContentWrapper, wixPalettes };